app/requirements.txt
flaskfastapi
uvicorn